@@ -9,16 +9,17 @@
         }
     ],
     "require": {
-        "php": "^7.4",
-        "maatwebsite/excel": "^3.0",
-        "illuminate/support": "^7.0",
-        "illuminate/database": "^7.0"
+        "php": "^8.1",
+        "maatwebsite/excel": "^3.1",
+        "illuminate/support": "^9.0",
+        "illuminate/database": "^9.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.5",
-        "orchestra/testbench": "^5.0",
+        "phpunit/phpunit": "^9.5.10",
+        "orchestra/testbench": "^7.7",
+        "pheromone/phpcs-security-audit": "^2.0",
         "mockery/mockery": "^1.3",
-        "phpstan/phpstan": "^0.12.20"
+        "phpstan/phpstan": "^1.9.0"
     },
     "autoload": {
         "psr-4": {
