@@ -47,12 +47,12 @@ class SqLite
      */
     public function hasTable(string $table): bool
     {
-         $bil = $this->connection
-            ->selectOne(
-                "select count(*) as bil from sqlite_master where type='table' AND name='" . $table ."'"
-            );
+        $bil = $this->connection
+           ->selectOne(
+               "select count(*) as bil from sqlite_master where type='table' AND name='" . $table ."'"
+           );
 
-         return (bool) $bil->bil;
+        return (bool) $bil->bil;
     }
 
     /**
